--- a/apps/comments/src/components/Comment.vue
+++ b/apps/comments/src/components/Comment.vue
@@ -110,7 +110,7 @@ import NcActions from '@nextcloud/vue/dist/Components/NcActions.js'
 import NcActionSeparator from '@nextcloud/vue/dist/Components/NcActionSeparator.js'
 import NcAvatar from '@nextcloud/vue/dist/Components/NcAvatar.js'
 import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
-import NcRichContenteditable from '@nextcloud/vue/dist/Components/NcRichContenteditable.js'
+const NcRichContenteditable = () => import('@nextcloud/vue/dist/Components/NcRichContenteditable')
 import RichEditorMixin from '@nextcloud/vue/dist/Mixins/richEditor.js'
 import ArrowRight from 'vue-material-design-icons/ArrowRight.vue'