--- a/apps/comments/src/components/Comment.vue
+++ b/apps/comments/src/components/Comment.vue
@@ -110,12 +110,14 @@ import NcActions from '@nextcloud/vue/dist/Components/NcActions.js'
 import NcActionSeparator from '@nextcloud/vue/dist/Components/NcActionSeparator.js'
 import NcAvatar from '@nextcloud/vue/dist/Components/NcAvatar.js'
 import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
-const NcRichContenteditable = () => import('@nextcloud/vue/dist/Components/NcRichContenteditable')
 import RichEditorMixin from '@nextcloud/vue/dist/Mixins/richEditor.js'
 import ArrowRight from 'vue-material-design-icons/ArrowRight.vue'
 
 import Moment from './Moment.vue'
 import CommentMixin from '../mixins/CommentMixin.js'
+
+// Dynamic loading
+const NcRichContenteditable = () => import('@nextcloud/vue/dist/Components/NcRichContenteditable.js')
 
 export default {
 	name: 'Comment',