Supercedes and closes #35432

This takes a much simpler approach for achieving something similar to the linked PR. Instead of lazy loading every component, webpack directly allows only splitting components used in more than `n` modules. By tweaking this parameter (3 seems to have great results), we can split the chunks efficiently achieving similar performance gains. This is also much less invasive and more semantically meaningful.

Some amount of async loading is still needed since apps greedily attempt to load a lot of things.

| Before                   | After                    |
|--------------------------|--------------------------|
|  ![wp_before](https://github.com/nextcloud/server/assets/10709794/01944f13-e18b-413c-ac5e-0d38e74f3bec) | ![wp_after](https://github.com/nextcloud/server/assets/10709794/b1c81392-7ea2-4335-9144-9086c76b483b) |

Files app (no cache):

| Before                   | After                    |
|--------------------------|--------------------------|
| 135 requests             | 138 requests             |
| 5.3 MB transferred       | 3.7 MB transferred       |
| 21.0 MB resources        | 12.2 MB resources        |
| DOMContentLoaded: 4.19 s | DOMContentLoaded: 3.02 s |

Files app (with cache):

| Before                   | After                    |
|--------------------------|--------------------------|
| 133 requests             | 133 requests             |
| 72.4 kB transferred      | 72.4 kB transferred      |
| 21.0 MB resources        | 12.2 MB resources        |
| DOMContentLoaded: 2.60 s | DOMContentLoaded: 1.39 s |

Memories (no cache):

| Before                   | After                    |
|--------------------------|--------------------------|
| 83 requests              | 83 requests              |
| 5.2 MB transferred       | 3.3 MB transferred       |
| 21.8 MB resources        | 12.3 MB resources        |
| DOMContentLoaded: 3.15 s | DOMContentLoaded: 2.25 s |

Video of load performance (before on top, after on bottom)
https://github.com/nextcloud/server/assets/10709794/5ab1144e-c811-4b8d-9798-463982b46cb7