@pulsejet 2023-05-17
Just noticed something funny: this is actually also a _bug fix_ at certain points. For instance, we explicitly _want_ to do lazy loading (before this PR) at this line for `caniuse-lite`. But as is evident from the bundle, the library was present in `core-common.js` all along, so the lazy loading never really worked.
https://github.com/nextcloud/server/blob/fa3bbb26538f8b56b2ab91dcfe3972a27a550d19/core/src/views/UnsupportedBrowser.vue#L125-L130