@pulsejet 2023-05-21
Short before (top) and after (bottom) video for this (warm cache for both):

https://github.com/nextcloud/server/assets/10709794/5ab1144e-c811-4b8d-9798-463982b46cb7