@pulsejet 2023-05-23
Thanks @kesselb. Just curious: doesn't the bot do that usually or I should be pushing compiled assets to the PR?