@skjnldsv 2023-05-24
> Bump? I really don't understand what keeps blocking this...

27 being released and master being in feature freeze.
This is now branched-off and we can resume