@skjnldsv 2023-05-24
> Thanks @kesselb. Just curious: doesn't the bot do that usually or I should be pushing compiled assets to the PR?

the bot needs to be called for it :)
`/compile amend /`