@major-mayer 2024-02-13
Now that I finally upgraded my instance to 28, I want to say thank you for this PR again @pulsejet 
This increases the performance noticeably!